WIFILocator:: New JADE Templates for Locations and Reviews
This commit adds several new JADE templates for locations and reviews.
The templates all extend the common layout temple, although there
still appears to be a small bug in that template that is pushing
the header over some of the content. Still need to add a
template for the site's about page.